fix rounding value and determinism by simplifying ranking function solution

+++ include/termination/ranking_and_invariant_validator.h
@@ -204,7 +204,7 @@ private:
         const std::vector<SupportingInvariant>& supporting_invariants,
         const LassoProgram& lasso,
         std::shared_ptr<SMTSolver> solver,
-        double delta,
+        int64_t delta,
         std::map<std::string, double>& counterexample);
 };
